refactor: extract functions into province-utils to simplify the mocks
