+++ lib/provinces.ts
@@ -31,37 +31,11 @@ export type ProvincePath = {
   };
 };
 
-export const getProvincesPaths = (): ProvincePath[] => {
-  const provincees = provinces as unknown as Province[];
-
-  return provincees.map((province) => {
-    const provinceSlug = province.slug;
-
-    return {
-      params: { provinceSlug },
-    };
-  });
-};
-
 export type ContactPath = {
   params: {
     provinceSlug: string;
     contactSlug: string;
   };
-};
-
-export const getContactsPaths = (): ContactPath[] => {
-  const contactsPaths: ContactPath[] = [];
-  provinces.forEach((province) => {
-    province.data.forEach((contact: Contact) => {
-      // eslint-disable-next-line @typescript-eslint/no-unsafe-argument
-      const provinceSlug = province.slug;
-      contactsPaths.push({
-        params: { provinceSlug, contactSlug: contact.slug },
-      });
-    });
-  });
-  return contactsPaths;
 };
 
 export default provinces as unknown as Provinces;
